Ready for main plot
data is in correct format.  It would be nice to find another quantitative variable, but I dont think I will.  Look at example Bokeh code.  Topics were unique and standardized to text_grouping.py is not needed and will not be finished.  Main file is complete to the actual bokeh creation.